+++ src/lib/server/db/migrations/0001_military_virginia_dare.sql
@@ -1,3 +1,0 @@
-ALTER TABLE "media" DROP CONSTRAINT "media_submission_id_submission_id_fk";
---> statement-breakpoint
-ALTER TABLE "media" DROP COLUMN "submission_id";